bfloat16

+++ DistributedSim/train_node.py
@@ -137,7 +137,7 @@ class TrainNode:
                 output = self.model(x_batch).transpose(1, 2)
                 loss = self.criterion(output, y_batch)
             else:
-                with torch.autocast(device_type=self.config.device_type, dtype=torch.float16):
+                with torch.autocast(device_type=self.config.device_type, dtype=torch.bfloat16):
                     output = self.model(x_batch).transpose(1, 2)
                     loss = self.criterion(output, y_batch)
 
@@ -194,7 +194,7 @@ class TrainNode:
                             output = self.model(x_batch).transpose(1, 2)
                             loss = self.criterion(output, y_batch)
                         else:
-                            with torch.autocast(device_type=self.config.device_type, dtype=torch.float16):
+                            with torch.autocast(device_type=self.config.device_type, dtype=torch.bfloat16):
                                 output = self.model(x_batch).transpose(1, 2)
                                 loss = self.criterion(output, y_batch)
